Add ready method to Trackers class
We need a way to know when the Tracker has finished loading the lists,
so that we know it's safe to access its corresponding state. Let's add
a `ready` method that works in a similar way to the `ready` functions
elsewhere in the extension code (by returning a Promise that resolves
once the list(s) is/are ready).